Move ratio_variance to pyquil/experiment/_result.py
